# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.ChangeCurPasswordForm:
    properties:
      new_password:
        maxLength: 32
        minLength: 4
        type: string
      old_password:
        maxLength: 32
        minLength: 4
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  admin.LoginPayload:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      route_names:
        items:
          type: string
        type: array
      token:
        type: string
      username:
        type: string
    type: object
...
  admin.RegisterForm:
    properties:
      confirm_password:
        maxLength: 32
        minLength: 4
        type: string
      email:
        description: validate:"required,email"
        type: string
      password:
        maxLength: 32
        minLength: 4
        type: string
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - confirm_password
    - password
    - username
    type: object
...
  admin.UpdateCurrentForm:
    properties:
      email:
        maxLength: 128
        type: string
      nickname:
        maxLength: 128
        type: string
    type: object
  admin.UserForm:
    properties:
      avatar:
        type: string
      email:
        description: validate:"required,email" email
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        description: Password string           `json:"password" validate:"required,gte=4,lte=20"`
        type: string
      remark:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.StatusCode'
        minimum: 0
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - group_id
    - status
    - username
    type: object
  admin.UserOauthItem:
    properties:
      op:
        type: string
      status:
        type: integer
    type: object
  admin.UserPasswordForm:
    properties:
      id:
        type: integer
      password:
        maxLength: 32
        minLength: 4
        type: string
    required:
    - id
    - password
    type: object
...
  model.StatusCode:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - COMMON_STATUS_ENABLE
    - COMMON_STATUS_DISABLED
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        type: string
      remark:
        type: string
      status:
        $ref: '#/definitions/model.StatusCode'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.User'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: API
  version: "1.0"
paths:
...
  /admin/my/groupUsers:
    post:
      description: Authenticated users receive only the group and user entries allowed
        for personal sharing.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: List shareable users and groups
      tags:
      - User
...
  /admin/user/changeCurPwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current user password change form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ChangeCurPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/changePwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: User password update form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/current:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.LoginPayload'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: User deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/groupUsers:
    post:
      description: Admin-only group and user picker data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: List users grouped by group
      tags:
      - User
  /admin/user/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for the user list
        in: query
        name: page
        type: integer
      - description: Number of users per page
        in: query
        name: page_size
        type: integer
      - description: Username filter
        in: query
        name: username
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/myOauth:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/admin.UserOauthItem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/register:
    post:
      consumes:
      - application/json
      description: Public registration route; the server decides whether the new account
        is enabled immediately.
      parameters:
      - description: User registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a user
      tags:
      - User
  /admin/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: User update form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/user/updateCurrent:
    post:
      consumes:
      - application/json
      description: Authenticated users may update only their nickname and email.
      parameters:
      - description: Current user profile update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateCurrentForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Update the current user profile
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"